+++ rollup.config.js
@@ -1,8 +1,9 @@
 import babel from 'rollup-plugin-babel';
+import commonjs from 'rollup-plugin-commonjs';
 import uglify from 'rollup-plugin-uglify';
 import { minify } from 'uglify-es';
+import nodeResolve from 'rollup-plugin-node-resolve';
 
-const format = 'es';
 const common = {
   external: ['react', 'prop-types', 'immutable', 'redux'],
   plugins: [
@@ -11,37 +12,61 @@ const common = {
       babelrc: false,
       presets: [
         [
-          '@babel/preset-stage-2',
+          '@babel/preset-env',
           {
-            loose: true,
-            useBuiltIns: true,
-            decoratorsLegacy: true,
+            targets: 'last 2 version',
           },
         ],
         '@babel/preset-flow',
       ],
+      plugins: ['@babel/plugin-proposal-class-properties'],
     }),
     uglify({}, minify),
   ],
 };
 
 export default [
+  // ES modules
+
   {
+    ...common,
     input: 'source/index.js',
     output: {
       name: 'reform-redux',
-      file: 'dist/reform-redux.js',
-      format,
+      file: 'reform-redux.es.js',
+      format: 'es',
     },
-    ...common,
   },
   {
+    ...common,
     input: 'source/immutable.js',
     output: {
       name: 'reform-redux',
-      file: 'dist/immutable.js',
-      format,
+      file: 'immutable.es.js',
+      format: 'es',
     },
+  },
+
+  // CommonJS
+
+  {
     ...common,
+    input: 'source/index.js',
+    output: {
+      name: 'reform-redux',
+      file: 'reform-redux.js',
+      format: 'cjs',
+    },
+    plugins: [...common.plugins, nodeResolve(), commonjs()],
+  },
+  {
+    ...common,
+    input: 'source/immutable.js',
+    output: {
+      name: 'reform-redux',
+      file: 'immutable.js',
+      format: 'cjs',
+    },
+    plugins: [...common.plugins, nodeResolve(), commonjs()],
   },
 ];